Move artwork cards into GalleryView
